# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FluxDB API",
    "version": "0.1.0",
    "description": "Single-node time-series database. JSON timestamps and i64 fields use decimal strings. The UI Developer resources page documents examples, SQL support and limitations."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8086"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/databases": {
      "get": {
        "summary": "List databases",
        "operationId": "get__api_v1_databases",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Database not found"
          },
          "413": {
            "description": "Body exceeds 2 MiB"
          },
          "422": {
            "description": "Invalid JSON shape"
          },
          "500": {
            "description": "Storage or internal error"
          }
        }
      }
    },
    "/api/v1/stats": {
      "get": {
        "summary": "Storage statistics",
        "operationId": "get__api_v1_stats",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Database not found"
          },
          "413": {
            "description": "Body exceeds 2 MiB"
          },
          "422": {
            "description": "Invalid JSON shape"
          },
          "500": {
            "description": "Storage or internal error"
          }
        }
      }
    },
...
    "/api/v1/databases/{name}": {
      "post": {
        "summary": "Create database",
        "operationId": "post__api_v1_databases_name",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Database not found"
          },
          "413": {
            "description": "Body exceeds 2 MiB"
          },
          "422": {
            "description": "Invalid JSON shape"
          },
          "500": {
            "description": "Storage or internal error"
          },
          "409": {
            "description": "Database exists"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ]
      },
      "delete": {
        "summary": "Permanently delete database",
        "operationId": "delete__api_v1_databases_name",
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Database not found"
          },
          "413": {
            "description": "Body exceeds 2 MiB"
          },
          "422": {
            "description": "Invalid JSON shape"
          },
          "500": {
            "description": "Storage or internal error"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ]
      }
    },
...
    "/api/v1/databases/{name}/query": {
      "post": {
        "summary": "Execute one SELECT in the supported SQL subset",
        "operationId": "post__api_v1_databases_name_query",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Database not found"
          },
          "413": {
            "description": "Body exceeds 2 MiB"
          },
          "422": {
            "description": "Invalid JSON shape"
          },
          "500": {
            "description": "Storage or internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "query"
                ],
                "properties": {
                  "query": {
                    "type": "string",
                    "maxLength": 32768
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}